Generate stub services (#3531)
* Update gulpfile with new projects and autorest 2.0 args

* Gen code for EventHub

* Gen code for ServiceFabric

* Gen code for NotificationHubs

* Gen code for AnalysisServices

* Gen code for Billing

* Gen code for CognitiveServices

* Gen code for Consumption

* Gen code for CustomerInsights

* Gen code for DevTestLab

* Gen code for IoTHub aka Devices

* Gen code for Logic

* Gen code for MachineLearning

* Remove unused project from Gulpfile

* Gen code for Relay

* Gen code for StorSimple

* Gen code for monitor management and data plane

* Bring gulpfile back to AutoRest 1.x conventions

* Update gitignore

* MonitorManager extends Manager<IMonitorManagementClient>

* Add IBeta to managers

* Remove sync methods

* Add AssemblyInfo.cs files

* Remove comments from auto-generated .csproj

+++ src/ResourceManagement/gulpfile.js
@@ -1,3 +1,4 @@
+var path = require('path');
 var gulp = require('gulp');
 var args = require('yargs').argv;
 var colors = require('colors');
@@ -11,6 +12,149 @@ var mappings = {
         'package': 'Microsoft.Azure.Management.Compute.Fluent',
         'composite': true,
         'args': '-FT 1'
+    },
+    'eventhub': {
+        'dir': 'EventHub',
+        'source': 'arm-eventhub/2015-08-01/swagger/EventHub.json',
+        'package': 'Microsoft.Azure.Management.EventHub.Fluent',
+        'args': '-FT 1'
+    },
+    'servicefabric': {
+        'dir': 'ServiceFabric',
+        'source': 'arm-servicefabric/2016-09-01/swagger/servicefabric.json',
+        'package': 'Microsoft.Azure.Management.ServiceFabric.Fluent',
+        'args': '-FT 1'
+    },
+    'notificationhubs': {
+        'dir': 'NotificationHubs',
+        'source': 'arm-notificationhubs/2017-04-01/swagger/notificationhubs.json',
+        'package': 'Microsoft.Azure.Management.NotificationHubs.Fluent',
+        'args': '-FT 1'
+    },
+    'analysisservices': {
+        'dir': 'AnalysisServices',
+        'source': 'arm-analysisservices/2016-05-16/swagger/analysisservices.json',
+        'package': 'Microsoft.Azure.Management.AnalysisServices.Fluent',
+        'args': '-FT 1'
+    },
+    'automation': {
+        'dir': 'Automation',
+        'source': 'arm-automation/compositeAutomation.json',
+        'package': 'Microsoft.Azure.Management.Automation.Fluent',
+        'args': '-FT 1',
+        'modeler': 'CompositeSwagger'
+    },
+    'billing': {
+        'dir': 'Billing',
+        'source': 'arm-billing/2017-04-24-preview/swagger/billing.json',
+        'package': 'Microsoft.Azure.Management.Billing.Fluent',
+        'args': '-FT 1'
+    },
+    'cognitiveservices': {
+        'dir': 'CognitiveServices',
+        'source': 'arm-cognitiveservices/2017-04-18/swagger/cognitiveservices.json',
+        'package': 'Microsoft.Azure.Management.CognitiveServices.Fluent',
+        'args': '-FT 1'
+    },
+    'consumption': {
+        'dir': 'Consumption',
+        'source': 'arm-consumption/2017-04-24-preview/swagger/consumption.json',
+        'package': 'Microsoft.Azure.Management.Consumption.Fluent',
+        'args': '-FT 1'
+    },
+    'customerinsights': {
+        'dir': 'CustomerInsights',
+        'source': 'arm-customer-insights/2017-04-26/swagger/customer-insights.json',
+        'package': 'Microsoft.Azure.Management.CustomerInsights.Fluent',
+        'args': '-FT 1'
+    },
+    'devtestlab': {
+        'dir': 'DevTestLab',
+        'source': 'arm-devtestlabs/2016-05-15/swagger/DTL.json',
+        'package': 'Microsoft.Azure.Management.DevTestLab.Fluent',
+        'args': '-FT 1'
+    },
+    'insights': {
+        'dir': 'Insights',
+        'source': 'arm-insights/compositeInsightsManagementClient.json',
+        'package': 'Microsoft.Azure.Management.Insights.Fluent',
+        'args': '-FT 1',
+        'modeler': 'CompositeSwagger'
+    },
+    'intune': {
+        'dir': 'Intune',
+        'source': 'arm-intune/2015-01-14-preview/swagger/intune.json',
+        'package': 'Microsoft.Azure.Management.Intune.Fluent',
+        'args': '-FT 1'
+    },
+    'iothub': {
+        'dir': 'Devices',
+        'source': 'arm-iothub/2017-01-19/swagger/iothub.json',
+        'package': 'Microsoft.Azure.Management.Devices.Fluent',
+        'args': '-FT 1'
+    },
+    'logic': {
+        'dir': 'Logic',
+        'source': 'arm-logic/2016-06-01/swagger/logic.json',
+        'package': 'Microsoft.Azure.Management.Logic.Fluent',
+        'args': '-FT 1'
+    },
+    'machinelearning': {
+        'dir': 'MachineLearning',
+        'source': 'arm-machinelearning/2017-01-01/swagger/webservices.json',
+        'package': 'Microsoft.Azure.Management.MachineLearning.Fluent',
+        'args': '-FT 1'
+    },
+    'operationalinsights': {
+        'dir': 'OperationalInsights',
+        'source': 'arm-operationalinsights/compositeOperationalInsights.json',
+        'package': 'Microsoft.Azure.Management.OperationalInsights.Fluent',
+        'args': '-FT 1',
+        'modeler': 'CompositeSwagger'
+    },
+    'powerbi': {
+        'dir': 'PowerBI',
+        'source': 'arm-powerbiembedded/2016-01-29/swagger/powerbiembedded.json',
+        'package': 'Microsoft.Azure.Management.PowerBI.Fluent',
+        'args': '-FT 1'
+    },
+    'recoveryservices': {
+        'dir': 'RecoveryServices',
+        'source': 'arm-recoveryservices/compositeRecoveryServicesClient.json',
+        'package': 'Microsoft.Azure.Management.RecoveryServices.Fluent',
+        'args': '-FT 1',
+        'modeler': 'CompositeSwagger'
+    },
+    'recoveryservicesbackup': {
+        'dir': 'RecoveryServices.Backup',
+        'source': 'arm-recoveryservicesbackup/2016-12-01/swagger/backupManagement.json',
+        'package': 'Microsoft.Azure.Management.RecoveryServices.Backup.Fluent',
+        'args': '-FT 1'
+    },
+    'recoveryservicessiterecovery': {
+        'dir': 'RecoveryServices.SiteRecovery',
+        'source': 'arm-recoveryservicessiterecovery/2016-08-10/swagger/service.json',
+        'package': 'Microsoft.Azure.Management.RecoveryServices.SiteRecovery.Fluent',
+        'args': '-FT 1'
+    },
+    'relay': {
+        'dir': 'Relay',
+        'source': 'arm-relay/2017-04-01/swagger/relay.json',
+        'package': 'Microsoft.Azure.Management.Relay.Fluent',
+        'args': '-FT 1'
+    },
+    'storsimple8000series': {
+        'dir': 'StorSimple',
+        'source': 'arm-storsimple8000series/2017-06-01/swagger/storsimple.json',
+        'package': 'Microsoft.Azure.Management.StorSimple.Fluent',
+        'args': '-FT 1'
+    },
+    'streamanalytics': {
+        'dir': 'StreamAnalytics',
+        'source': 'arm-streamanalytics/compositeStreamAnalytics.json',
+        'package': 'Microsoft.Azure.Management.StreamAnalytics.Fluent',
+        'args': '-FT 1',
+        'modeler': 'CompositeSwagger'
     },
     'storage': {
         'dir': 'Storage',
@@ -106,6 +250,20 @@ var mappings = {
         'package': 'Microsoft.Azure.Management.Fluent.ServiceBus',
         'args': '-FT 1'
     },
+    'monitor': {
+        'dir': 'Monitor',
+        'source': 'arm-monitor/compositeMonitorManagementClient.json',
+        'package': 'Microsoft.Azure.Management.Fluent.ServiceBus',
+        'args': '-FT 1 -ServiceName Monitor',
+        'composite': true
+    },
+    'monitor-dataplane': {
+        'dir': 'Monitor',
+        'source': 'monitor/compositeMonitorClient.json',
+        'package': 'Microsoft.Azure.Management.Fluent.ServiceBus',
+        'args': '-FT 1 -ServiceName Monitor',
+        'composite': true
+    },
     'containerregistry': {
         'dir': 'ContainerRegistry',
         'source': 'arm-containerregistry/2017-03-01/swagger/containerregistry.json',
@@ -120,12 +278,12 @@ var mappings = {
     }
 };
 
-gulp.task('default', function() {
+gulp.task('default', function () {
     console.log("Usage: gulp codegen [--spec-root <swagger specs root>] [--projects <project names>] [--autorest <autorest info>]\n");
     console.log("--spec-root");
     console.log("\tRoot location of Swagger API specs, default value is \"https://raw.githubusercontent.com/Azure/azure-rest-api-specs/master\"");
     console.log("--projects\n\tComma separated projects to regenerate, default is all. List of available project names:");
-    Object.keys(mappings).forEach(function(i) {
+    Object.keys(mappings).forEach(function (i) {
         console.log('\t' + i.magenta);
     });
     console.log("--autorest\n\tThe version of AutoRest. E.g. 1.0.1-20170222-2300-nightly, or the location of AutoRest repo, E.g. E:\\repo\\autorest");
@@ -140,28 +298,28 @@ if (args['autorest'] !== undefined) {
 }
 var autoRestExe;
 
-gulp.task('codegen', function(cb) {
+gulp.task('codegen', function (cb) {
     if (autoRestVersion.match(/[0-9]+\.[0-9]+\.[0-9]+.*/) ||
         autoRestVersion == 'latest') {
         autoRestExe = 'autorest ---version=' + autoRestVersion;
         handleInput(projects, cb);
     } else {
-        autoRestExe = autoRestVersion + "/" + GetAutoRestFolder() + "AutoRest.exe";
+        autoRestExe = path.join(autoRestVersion, "src/core/AutoRest/bin/Debug/netcoreapp1.0/AutoRest.dll");
         if (process.platform !== 'win32') {
-            autoRestExe = "mono " + autoRestExe;
+            autoRestExe = "dotnet " + autoRestExe;
         }
         handleInput(projects, cb);
     }
 
 });
 
-var handleInput = function(projects, cb) {
+var handleInput = function (projects, cb) {
     if (projects === undefined) {
-        Object.keys(mappings).forEach(function(proj) {
+        Object.keys(mappings).forEach(function (proj) {
             codegen(proj, cb);
         });
     } else {
-        projects.split(",").forEach(function(proj) {
+        projects.split(",").forEach(function (proj) {
             proj = proj.replace(/\ /g, '');
             if (mappings[proj] === undefined) {
                 console.error('Invalid project name "' + proj + '"!');
@@ -172,8 +330,10 @@ var handleInput = function(projects, cb) {
     }
 }
 
-var codegen = function(project, cb) {
-    deleteFolderRecursive(mappings[project].dir + '/Generated');
+var codegen = function (project, cb) {
+    if (!args['preserve']) {
+        deleteFolderRecursive(mappings[project].dir + '/Generated');
+    }
     console.log('Generating "' + project + '" from spec file ' + specRoot + '/' + mappings[project].source);
     var generator = 'Azure.CSharp.Fluent';
     if (mappings[project].fluent !== null && mappings[project].fluent === false) {
@@ -183,28 +343,29 @@ var codegen = function(project, cb) {
     if (mappings[project].composite !== null && mappings[project].composite === true) {
         modeler = 'CompositeSwagger';
     }
-    cmd = autoRestExe + ' -Modeler ' + modeler + 
-                        ' -CodeGenerator ' + generator + 
-                        ' -Namespace ' + mappings[project].package + 
-                        ' -Input ' + specRoot + '/' + mappings[project].source + 
-                        ' -outputDirectory ' + mappings[project].dir + '/Generated' + 
-                        ' -Header MICROSOFT_MIT' +
-                        ' -skipValidation true';
+    cmd = autoRestExe + ' -Modeler ' + modeler +
+        ' -CodeGenerator ' + generator +
+        ' -Namespace ' + mappings[project].package +
+        ' -Input ' + specRoot + '/' + mappings[project].source +
+        ' -outputDirectory ' + mappings[project].dir +
+        ' -Header MICROSOFT_MIT' +
+        ' -RegenerateManager true' +
+        ' -skipValidation true';
     if (mappings[project].args !== undefined) {
         cmd = cmd + ' ' + mappings[project].args;
     }
     console.log('Command: ' + cmd);
-    exec(cmd, function(err, stdout, stderr) {
+    exec(cmd, function (err, stdout, stderr) {
         console.log(stdout);
         console.error(stderr);
     });
 };
 
-var deleteFolderRecursive = function(path) {
-    if(fs.existsSync(path)) {
-        fs.readdirSync(path).forEach(function(file, index) {
+var deleteFolderRecursive = function (path) {
+    if (fs.existsSync(path)) {
+        fs.readdirSync(path).forEach(function (file, index) {
             var curPath = path + "/" + file;
-            if(fs.lstatSync(curPath).isDirectory()) { // recurse
+            if (fs.lstatSync(curPath).isDirectory()) { // recurse
                 deleteFolderRecursive(curPath);
             } else { // delete file
                 fs.unlinkSync(curPath);
@@ -213,21 +374,3 @@ var deleteFolderRecursive = function(path) {
         fs.rmdirSync(path);
     }
 };
-
-
-var isWindows = (process.platform.lastIndexOf('win') === 0);
-var isLinux= (process.platform.lastIndexOf('linux') === 0);
-var isMac = (process.platform.lastIndexOf('darwin') === 0);
-
-function GetAutoRestFolder() {
-  if (isWindows) {
-    return "src/core/AutoRest/bin/Debug/net451/win7-x64/";
-  }
-  if( isMac ) {
-    return "src/core/AutoRest/bin/Debug/net451/osx.10.11-x64/";
-  } 
-  if( isLinux ) { 
-    return "src/core/AutoRest/bin/Debug/net451/ubuntu.14.04-x64/"
-  }
-   throw new Error("Unknown platform?");
-}
